create term create school ,school filter page

+++ src/utils/axios.js
@@ -3,20 +3,37 @@ import { baseUrl } from "./constant";
 
 const Axios = axios.create({
     baseURL: baseUrl
-})
+});
+
+Axios.interceptors.request.use(
+    (config) => {
+        const token = localStorage.getItem("token");
+        if (token) {
+            config.headers.Authorization = `Bearer ${token}`;
+        }
+        return config;
+    },
+    (error) => Promise.reject(error)
+);
 
 Axios.interceptors.response.use(
-    response => {
-        return response
-    },
-    error => {
-        if (error.response.status === 401) {
-            localStorage.clear()
-            window.location.href('/')
+    (response) => {
+        if (response && response.data) {
+            return response;
+        } else {
+            return Promise.reject({ message: "No response received from server" });
         }
-        else if (error.response && error.response.data)
-            return Promise.reject(error.response.data)
-        else return Promise.reject({ message: "Something went wrong" })
+    },
+    (error) => {
+        if (error.response && error.response.status === 401) {
+            localStorage.clear();
+            window.location.href = "/";
+        } else if (error.response && error.response.data) {
+            return Promise.reject(error.response.data);
+        } else {
+            return Promise.reject({ message: "Something went wrong" });
+        }
     }
-)
-export default Axios
+);
+
+export default Axios;
